+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 0.1 (unreleased)
 ----------------
 
+- Add base class for workflow adaptation.
+  [sgeulette]
 - Initial release.
   [cedricmessiant]
-